requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --format requirements-txt
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via httpx
beautifulsoup4==4.14.3
    # via fundor333-github-io
bleach==6.3.0
    # via fundor333-github-io
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   typer
    #   weeknotebot
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
feedparser==6.0.12
    # via
    #   fundor333-github-io
    #   weeknotebot
fonttools==4.62.1
    # via matplotlib
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via ollama
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
kiwisolver==1.5.0
    # via matplotlib
lxml==6.0.2
    # via fundor333-github-io
markdown-it-py==4.0.0
    # via rich
marshmallow==3.26.2
    # via weeknotebot
matplotlib==3.10.8
    # via fundor333-github-io
mdurl==0.1.2
    # via markdown-it-py
numpy==2.4.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
ollama==0.6.1
    # via fundor333-github-io
packaging==26.0
    # via
    #   marshmallow
    #   matplotlib
pandas==3.0.2
    # via fundor333-github-io
pillow==12.2.0
    # via
    #   fundor333-github-io
    #   matplotlib
pydantic==2.12.5
    # via ollama
pydantic-core==2.41.5
    # via pydantic
pygments==2.20.0
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.2
    # via fundor333-github-io
python-frontmatter==1.1.0
    # via fundor333-github-io
pyyaml==6.0.3
    # via
    #   fundor333-github-io
    #   python-frontmatter
requests==2.33.1
    # via
    #   fundor333-github-io
    #   weeknotebot
rich==13.9.4
    # via
    #   fundor333-github-io
    #   typer
    #   weeknotebot
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
typer==0.24.1
    # via fundor333-github-io
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2026.1 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via pandas
urllib3==2.6.3
    # via requests
webencodings==0.5.1
    # via bleach
weeknotebot==1.10.0
    # via fundor333-github-io
